♻️ refactor: url폴더 상위에 const폴더 추가
