+++ src/common/const/index.ts
@@ -1,0 +1,1 @@
+export * from './resourceUrl';
