+++ src/components/Comment/index.tsx
@@ -1,4 +1,4 @@
-import { ICON_URL } from '@/common/resourceUrl';
+import { ICON_URL } from '@/common/const/';
 import clsx from 'clsx';
 import Image from 'next/image';
